+++ Plugins/iOS/FlurryUnityWrapper.m
@@ -72,7 +72,7 @@ static FlurryUnityWrapper *_sharedInstance;
     NSLog(@"Flurry session started");
     
     NSString* originName = @"unity-flurry-sdk";
-    NSString* originVersion = @"6.0.0";
+    NSString* originVersion = @"6.1.0";
     
     [Flurry addOrigin:originName withVersion:originVersion];