+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 * [Flurry Android SDK Release Notes](https://developer.yahoo.com/flurry/docs/releasenotes/android/)
 * [Flurry iOS SDK Release Notes](https://developer.yahoo.com/flurry/docs/releasenotes/ios/)
 
+## v6.1.0 (2023-02-02)
+
+#### Features
+
+* Upgrade Flurry Android SDK version to 14.1.0
+* Support GPP Consents APIs (Android only)
+
 # v6.0.0 (2022-10-31)
 
 #### Features